alta_data_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pydantic-settings
sqlalchemy>=2
asyncpg
alembic
passlib[bcrypt]
python-jose[cryptography]
structlog
celery[redis]
celery
redis
google-cloud-storage
google-cloud-documentai
google-cloud-speech
httpx
jinja2
email-validator
python-multipart
flower
kombu
